requirements.txt
typing  
langchain.llms  
langchain.prompts
transformers 
torch
pinecone


